refactor: Move aclients logic to cmd/samoyed-aclients
AClientsMain and its helpers move out of the src/direwolf package and
into the cmd/samoyed-aclients main package, following the pattern used
by other already-migrated commands (kissutil, axudp, log2gpx, ll2utm).

LINE_WIDTH and column_width move into cmd/samoyed-aclients/main.go too;
tnctest (moved to cmd/samoyed-tnctest separately) already keeps its own
local copies, so there's no longer any need to share this column
layout state via the direwolf package.
